gotoCtx should just take a MIR context at construction time
